Ensure IRC formatting doesn't result in an unhandled error
The old formatters were just pulled out of the old IRC backend. This
refactors them so that they should always return a useful string to send
even if fedmsg.meta falls over. This stops it from using the raw fedmsg
which the backend couldn't send since it got serialized as a dict.

Fixes #263

Signed-off-by: Jeremy Cline <jeremy@jcline.org>